refactoring not-found-screen
