+++ src/pages/not-found-screen/not-found-screen.module.css
@@ -1,0 +1,62 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+}
+
+.logo {
+  padding: 10px;
+}
+
+.main {
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  padding-top: 20px;
+}
+
+.title {
+  font-size: 5rem;
+  font-weight: bold;
+  margin: 0;
+}
+
+.message {
+  font-size: 1.5rem;
+  margin: 5px 0;
+}
+
+.pin {
+  font-size: 5rem;
+  margin-top: 10px;
+  animation: bounce 1s infinite;
+}
+
+.secondaryMessage {
+  font-size: 1.2rem;
+  margin-top: 10px;
+}
+
+.button {
+  background-color: #007bff;
+  color: #fff;
+  padding: 10px 20px;
+  border-radius: 5px;
+  text-decoration: none;
+  font-size: 1.2rem;
+  margin-top: 20px;
+  display: flex;
+  align-items: center;
+}
+
+@keyframes bounce {
+  0%,
+  100% {
+    transform: translateY(0);
+  }
+  50% {
+    transform: translateY(-20px);
+  }
+}
